Add event listeners for new-tweet
Adds event listeners to handle text input and response.

On input event into the textarea of the new-tweet section form, the character length is calculated and then subtracted from the value in output.

This also removes the default behaviour of submit event of the form. Instead, on submit the text area is cleared, the output reset to 140, and the textarea is given focus again.