Add createTweetElement and renderTweets
createTweetElement takes in an object with tweet data values and renders an html template populated with the values from the tweet object argument.

renderTweets takes an array/object of tweet objects and cycles through them calling createTweetElement on each tweet object. It then appends the generated tweet HTML to the tweet-log section.

If more modularity is needed, these functions could be generalised to take more arguments and be flexible in what the target is to append to.