Move event handlers composer to client.js
Moves all event handlers from composer-char-counter.js to client.js to introduce conditions to display a window alert in case of invalid lenth for the textarea form input on the submit event.

Event handlers that reset the newTweet elements are moved into the resolved promise block of the ajax POST request.