Refactor CSS for responsive design
Sets break point for 421px width which changes the nav, and main elements.

In the mobile version, the nav becomes transparent and each successive element is stacked vertically.

In the tablet and above versions, the nav is coloured and the header and .tweets sections are side by side.

Header changes to be sticky as well as some refactoring of the user-name and img elements to fit better in proportion to changing header dimensions.